requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
faiss-cpu==1.7.4
numpy==1.26.0
Pillow==10.0.1
tqdm==4.66.1
torch==2.0.1
transformers==4.34.0
googletrans==4.0.0-rc1
python-multipart==0.0.6
opencv-python==4.8.1.78 